requirements.txt
fastapi
openai
python-dotenv
requests
pydantic
pydantic-settings
pytest